requirements.txt
discord.py
asyncio
aiohttp
